# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farmácia Pai e Filha API",
        "description": "API para gerenciamento da Farmácia Pai e Filha",
        "version": "1.0.0"
    },
    "paths": {
...
        "/batches/toggle-status/{id}": {
            "patch": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Ativa ou Inativa um lote específico",
                "operationId": "4e7dcbb0776deb637f3592a01aa5cb03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status alterado com sucesso"
                    }
                }
            }
        },
...
        "/batches/update/{id}": {
            "put": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Atualiza os dados de um lote específico",
                "operationId": "5ff29b9245f1151937b448e06ce0895c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "batch_code": {
                                        "type": "string"
                                    },
                                    "quantity_now": {
                                        "type": "integer"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "cost_price": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lote atualizado"
                    },
                    "404": {
                        "description": "Lote não encontrado"
                    }
                }
            }
        },
        "/product/{id}/deactivate-batches": {
            "patch": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Desativa todos os lotes ativos de um produto",
                "operationId": "45979e945cde34a1a5a1bedf6ddc889a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lotes desativados com sucesso"
                    }
                }
            }
        },
...
    },
...
}